Refactor of style classes, using SCSS
- Removing SASS and using SCSS instead.
- Adding SCSS file for each component/page/template for easier maintainence.
- Removing components and styles not in use.
- Still need to remove bulma style library that is underneath and also test with CMS.
