--- src/components/Navbar.js
@@ -1,101 +1,0 @@
-import React from 'react'
-import { Link } from 'gatsby'
-// import github from '../img/github-icon.svg'
-// import logo from '../img/logo.png'
-
-const Navbar = class extends React.Component {
-
-  componentDidMount() {
-    // Get all "navbar-burger" elements
-   const $navbarBurgers = Array.prototype.slice.call(document.querySelectorAll('.navbar-burger'), 0);
-    // Check if there are any navbar burgers
-   if ($navbarBurgers.length > 0) {
-
-     // Add a click event on each of them
-     $navbarBurgers.forEach( el => {
-       el.addEventListener('click', () => {
-
-         // Get the target from the "data-target" attribute
-         const target = el.dataset.target;
-         const $target = document.getElementById(target);
-
-         // Toggle the "is-active" class on both the "navbar-burger" and the "navbar-menu"
-         el.classList.toggle('is-active');
-         $target.classList.toggle('is-active');
-
-       });
-     });
-   }
- }
-
- render() {
-   return (
-     <nav className="nav-wrapper">
-      <div className="nav">
-          <Link className="nav-title" to="/">SAP Fiori</Link>
-          <div className="navbar">
-            <Link className="nav-item" to="/design/guidelines">
-              Design
-          </Link>
-          <Link className="nav-item" to="/develop/guidelines">
-            Develop
-          </Link>
-          <Link className="nav-item" to="/getstarted">
-            Get Started
-          </Link>
-          <Link className="nav-item" to="/contact/examples">
-            Contribute
-          </Link>
-          <div className="nav-search"></div>
-          </div>
-      </div>
-     </nav>
-
-  // <nav className="navbar is-transparent" role="navigation" aria-label="main-navigation">
-  //   <div className="container">
-  //     <div className="navbar-brand">
-  //       <Link to="/" className="navbar-item" title="Logo">
-  //         <img src={logo} alt="SAP Fiori" style={{ width: '60px' }} />
-  //         SAP Fiori
-  //       </Link>
-  //       {/* Hamburger menu */}
-  //       <div className="navbar-burger burger" data-target="navMenu">
-  //         <span></span>
-  //         <span></span>
-  //         <span></span>
-  //       </div>
-  //     </div>
-  //     <div id="navMenu" className="navbar-menu">
-  //     <div className="navbar-start has-text-centered">
-  //       <Link className="navbar-item" to="/about">
-  //         Design
-  //       </Link>
-  //       <Link className="navbar-item" to="/products">
-  //         Develop
-  //       </Link>
-  //       <Link className="navbar-item" to="/getstarted">
-  //         Get Started
-  //       </Link>
-  //       <Link className="navbar-item" to="/contact/examples">
-  //         Contribute
-  //       </Link>
-  //     </div>
-  //     <div className="navbar-end has-text-centered">
-  //       <a
-  //         className="navbar-item"
-  //         href="https://github.com/devinea/mvx19-cms"
-  //         target="_blank"
-  //         rel="noopener noreferrer"
-  //       >
-  //         <span className="icon">
-  //           <img src={github} alt="Github" />
-  //         </span>
-  //       </a>
-  //     </div>
-  //     </div>
-  //   </div>
-  // </nav>
-  )}
-}
-
-export default Navbar
