+++ src/components/all.sass
@@ -1,257 +1,2 @@
-@import "~bulma/sass/utilities/initial-variables"
-
-// Config vars
-// Kalidi logo brand color: #FD461E
-$kaldi-red: #D64000
-$kaldi-red-invert: #fff
-
-$primary: $kaldi-red
-$primary-invert: $kaldi-red-invert
-$body-color: #333
-
-@font-face
-  font-family: '72-Regular';
-  src: url('../fonts/72-Regular.ttf') format('truetype');
-  font-weight: normal;
-  font-style: normal;
-
-@font-face
-  font-family: '72-Light';
-  src: url('../fonts/72-Light.ttf') format('truetype');
-  font-weight: normal;
-  font-style: normal;
-
-@font-face
-  font-family: '72-Bold';
-  src: url('../fonts/72-Bold.ttf') format('truetype');
-  font-weight: normal;
-  font-style: normal;
-
-.main-content
-  padding-top: 70px;
-  width: 100%;
-  float: left;
-  background-color: #F5F5F5;
-
-.trending
-  width: 1440px;
-  height: 750px;
-  background-image:  url("../img/trending-back.png");
-  background-repeat: no-repeat;
-  background-position: center top;
-  background-color: white;
-  margin-left: auto;
-  margin-right: auto;
-  position: relative;
-
-  .trending-footer
-    position: absolute;
-    bottom: 0px;
-    background-color: #242939;
-    width: 100%;
-    height: 116px;
-
-  .trending-content
-    font-family: '72-Light';
-    padding-left: 115px;
-    padding-top: 160px;
-
-    .trending-header
-      font-size: 30px;
-      width: 475px;
-      color: #818181;
-
-    .trending-search
-      padding-top: 20px;
-      input
-        width: 490px;
-        height: 45px;
-        padding-left: 10px;
-        padding-right: 50px;
-        border-radius: 15px;
-        border: none;
-        box-shadow: 1px 1px 10px 1px lightgray;
-        background-image: url("../img/search.png");
-        background-repeat: no-repeat;
-        background-position: right 20px center;
-      input:focus
-        outline-width: 0;
-
-    .trending-suggestions
-      color: #ACACAC;
-      margin-top: 35px;
-      font-size: 16px;
-      div
-        line-height: 26px;
-      .trending-title
-        font-weight: bold;
-  
-.whatsnew
-  width: 1440px;
-  background-color: white;
-  margin-left: auto;
-  margin-right: auto;
-  padding-left: 0px;
-  padding: 40px 0 20px 0 !important;
-
-  .whatsnew-wrapper 
-    float: left;
-    width: 100%;
-    background-color: white;
-    .container
-      padding-left: 40px;
-
-    .content
-      .whatsnew-header
-        clear: both;
-        width: 100%;
-        height: 50px;
-        .whatsnew-title
-          font-family: '72-Bold';
-          font-size: 35px;
-          float: left;
-        .whatsnew-options
-          float: right;
-          div
-            float: left;
-            font-family: '72-Light';
-            font-size: 28px;
-            color: #A1A1A1;
-            cursor: pointer;
-          div:not(:last-child)   
-            padding-right: 150px;   
-          div.selected
-            color: #5C5C5C;
-            cursor: default;
-    .new-lhs 
-      width: 385px;
-      float: left;
-      background-color: white;
-      .new-lhs-title
-        font-family: '72-Light';
-        font-size: 24px;
-        color: #7E7E7E;
-        width: 280px;
-        margin-bottom: 100px;
-      .new-lhs-option
-        font-family: '72-Light';
-        font-size: 24px;
-        color: #A1A1A1;
-        line-height: 55px;
-        cursor: pointer;
-      .new-lhs-option.selected
-        color: #5C5C5C;
-        cursor: default;
-    .new-rhs
-      float: right;
-      width: calc(100% - 385px);
-      padding-bottom: 40px;
-
-      .whatsnew-items
-        padding: 30px 30px 30px 160px;
-        .newitem-title
-          font-family: '72-Bold';
-          font-size: 24px;
-          color: #5C5C5C;
-        .newitem-details
-          font-family: '72-Light';
-          font-size: 18px;
-          color: #A1A1A1;
-          float: left;
-          width: calc(100% - 150px);
-        .newitem-more
-          float: right;
-          font-family: '72-Bold';
-          font-size: 14px;
-          text-transform: uppercase;
-          color: #A1A1A1;
-          width: 150px;
-          text-align: right;
-          cursor: pointer;
-        .newitem-date
-          clear: both;
-          font-family: '72-Light';
-          font-size: 14px;
-          color: #A1A1A1;  
-          padding-top: 20px;
-        background-repeat: no-repeat;
-        &:hover
-          box-shadow: 1px 1px 10px 1px lightgray;   
-        &:nth-child(1) 
-          background-image: url("../img/sketch.png");
-          background-size: 75px;
-          background-position: 50px 30px;
-        &:nth-child(2) 
-          background-image: url("../img/ui5.jpg");
-          background-size: 125px;
-          background-position: 25px 30px;          
-        &:nth-child(3) 
-          background-image: url("../img/github.svg");
-          background-size: 75px;
-          background-position: 50px 30px;     
-        &:nth-child(4) 
-          background-image: url("../img/theming.png");
-          background-size: 75px;
-          background-position: 50px 30px;      
-        &:nth-child(5) 
-          background-image: url("../img/design.png");
-          background-size: 125px;
-          background-position: 25px 30px;                          
-        
-
-
-.content .taglist
-  list-style: none
-  margin-bottom: 0
-  margin-left: 0
-  margin-right: 1.5rem
-  margin-top: 1.5rem
-  display: flex
-  flex-wrap: wrap
-  justify-content: left
-  align-items: center
-  li
-    padding: 0 2rem 1rem 0
-    margin-bottom: 1.5rem
-    margin-top: 0
-
-// Helper Classes
-.full-width-image-container
-  width: 100vw
-  height: 400px
-  position: relative
-  left: 50%
-  right: 50%
-  margin: 5em -50vw
-  background-size: cover
-  background-position: bottom
-  display: flex
-  justify-content: center
-  align-items: center
-
-.margin-top-0
-  margin-top: 0 !important
-
-
-.blog
-  padding: 40px 0 0 40px !important;
-  width: 1440px;
-  background-color: white;
-  margin-left: auto;
-  margin-right: auto;
-
-.about
-  padding: 40px 0 0 40px !important;
-  width: 1440px;
-  background-color: white;
-  margin-left: auto;
-  margin-right: auto;
-
-.contact
-  padding: 40px 0 0 40px !important;
-  width: 1440px;
-  background-color: white;
-  margin-left: auto;
-  margin-right: auto;  
 
 @import "~bulma"
