+++ src/components/layout.scss
@@ -1,0 +1,27 @@
+@font-face {
+    font-family: '72-Regular';
+    src: url('../fonts/72-Regular.ttf') format('truetype');
+    font-weight: normal;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: '72-Light';
+    src: url('../fonts/72-Light.ttf') format('truetype');
+    font-weight: normal;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: '72-Bold';
+    src: url('../fonts/72-Bold.ttf') format('truetype');
+    font-weight: normal;
+    font-style: normal;
+}
+
+.main-content {
+  padding-top: 70px;
+  width: 100%;
+  float: left;
+  background-color: #F5F5F5;
+}
