+++ src/components/nav-bar/nav-bar.js
@@ -1,0 +1,56 @@
+import React from 'react'
+import { Link } from 'gatsby'
+import './nav-bar.scss'
+
+/**
+ * Generates the navigation bar
+ */
+const Navbar = class extends React.Component {
+
+  componentDidMount() {
+    // Get all "navbar-burger" elements
+   const $navbarBurgers = Array.prototype.slice.call(document.querySelectorAll('.navbar-burger'), 0);
+    // Check if there are any navbar burgers
+   if ($navbarBurgers.length > 0) {
+
+     // Add a click event on each of them
+     $navbarBurgers.forEach( el => {
+       el.addEventListener('click', () => {
+         // Get the target from the "data-target" attribute
+         const target = el.dataset.target;
+         const $target = document.getElementById(target);
+
+         // Toggle the "is-active" class on both the "navbar-burger" and the "navbar-menu"
+         el.classList.toggle('is-active');
+         $target.classList.toggle('is-active');
+       });
+     });
+   }
+ }
+
+ render() {
+   return (
+     <nav className="nav-wrapper">
+      <div className="nav">
+          <Link className="nav-title" to="/">SAP Fiori</Link>
+          <div className="navbar">
+            <Link className="nav-item" to="/design/guidelines">
+              Design
+          </Link>
+          <Link className="nav-item" to="/develop/guidelines">
+            Develop
+          </Link>
+          <Link className="nav-item" to="/getstarted">
+            Get Started
+          </Link>
+          <Link className="nav-item" to="/contact/examples">
+            Contribute
+          </Link>
+          <div className="nav-search"></div>
+          </div>
+      </div>
+     </nav>
+  )}
+}
+
+export default Navbar
