--- src/components/nav.sass
@@ -1,42 +1,0 @@
-.nav-wrapper 
-  position: fixed;
-  z-index: 100;
-  width: 100vw;
-
-.nav
-  width: 1440px;
-  height: 70px;
-  margin-left: auto;
-  margin-right: auto;
-  line-height: 70px;
-  background-color: white;
-  border-bottom: 1px #eaeaea solid;
-
-  .nav-title
-    float: left;
-    padding-left: 80px;
-    font-weight: bold;
-    background-image: url("../img/logo.png");
-    background-position: 30px center;
-    background-repeat: no-repeat;
-    background-size: 40px;
-    color: black;
-
-  .navbar
-    float: right;
-    height: 69px;
-
-    .nav-item
-      width: 150px;
-      text-align: center;
-      font-weight: bold;
-      color: #515151;
-
-    .nav-search
-      width: 20px;
-      background-image: url("../img/search.png");
-      margin-right: 30px;
-      background-position: center center;
-      background-repeat: no-repeat;
-      cursor: pointer;
-
