+++ src/pages/contact/contact.scss
@@ -1,0 +1,7 @@
+.contact {
+  padding: 40px 0 0 40px !important;
+  width: 1440px;
+  background-color: white;
+  margin-left: auto;
+  margin-right: auto;  
+}
