+++ src/pages/contact/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { navigate } from "gatsby-link";
 import Layout from '../../components/Layout'
+import './contact.scss'
 
 function encode(data) {
   return Object.keys(data)
