+++ src/templates/blog-post/blog-post.scss
@@ -1,0 +1,7 @@
+.blog {
+  padding: 40px 0 0 40px !important;
+  width: 1440px;
+  background-color: white;
+  margin-left: auto;
+  margin-right: auto;
+}
