+++ src/templates/design-guideline-post/design-guideline-post.js
@@ -3,10 +3,11 @@ import PropTypes from 'prop-types'
 import { kebabCase } from 'lodash'
 import Helmet from 'react-helmet'
 import { graphql, Link } from 'gatsby'
-import Layout from '../components/Layout'
-import Content, { HTMLContent } from '../components/Content'
+import Layout from '../../components/Layout'
+import Content, { HTMLContent } from '../../components/Content'
+import './design-guideline-post.scss'
 
-export const LearningPostTemplate = ({
+export const DesignGuidelinePostTemplate = ({
   content,
   contentComponent,
   description,
@@ -46,7 +47,7 @@ export const LearningPostTemplate = ({
   )
 }
 
-LearningPostTemplate.propTypes = {
+DesignGuidelinePostTemplate.propTypes = {
   content: PropTypes.node.isRequired,
   contentComponent: PropTypes.func,
   description: PropTypes.string,
@@ -54,12 +55,12 @@ LearningPostTemplate.propTypes = {
   helmet: PropTypes.object,
 }
 
-const LearningPost = ({ data }) => {
+const DesignGuidelinePost = ({ data }) => {
   const { markdownRemark: post } = data
 
   return (
     <Layout>
-      <LearningPostTemplate
+      <DesignGuidelinePostTemplate
         content={post.html}
         contentComponent={HTMLContent}
         description={post.frontmatter.description}
@@ -78,16 +79,16 @@ const LearningPost = ({ data }) => {
   )
 }
 
-LearningPost.propTypes = {
+DesignGuidelinePost.propTypes = {
   data: PropTypes.shape({
     markdownRemark: PropTypes.object,
   }),
 }
 
-export default LearningPost
+export default DesignGuidelinePost
 
 export const pageQuery = graphql`
-  query LearningPostByID($id: String!) {
+  query DesignGuidelinePostByID($id: String!) {
     markdownRemark(id: { eq: $id }) {
       id
       html
