+++ src/templates/developer-guideline-post/developer-guideline-post.js
@@ -3,10 +3,11 @@ import PropTypes from 'prop-types'
 import { kebabCase } from 'lodash'
 import Helmet from 'react-helmet'
 import { graphql, Link } from 'gatsby'
-import Layout from '../components/Layout'
-import Content, { HTMLContent } from '../components/Content'
+import Layout from '../../components/Layout'
+import Content, { HTMLContent } from '../../components/Content'
+import './developer-guideline-post.scss'
 
-export const DesignGuidelinePostTemplate = ({
+export const DeveloperGuidelinePostTemplate = ({
   content,
   contentComponent,
   description,
@@ -46,7 +47,7 @@ export const DesignGuidelinePostTemplate = ({
   )
 }
 
-DesignGuidelinePostTemplate.propTypes = {
+DeveloperGuidelinePostTemplate.propTypes = {
   content: PropTypes.node.isRequired,
   contentComponent: PropTypes.func,
   description: PropTypes.string,
@@ -54,12 +55,12 @@ DesignGuidelinePostTemplate.propTypes = {
   helmet: PropTypes.object,
 }
 
-const DesignGuidelinePost = ({ data }) => {
+const DeveloperGuidelinePost = ({ data }) => {
   const { markdownRemark: post } = data
 
   return (
     <Layout>
-      <DesignGuidelinePostTemplate
+      <DeveloperGuidelinePostTemplate
         content={post.html}
         contentComponent={HTMLContent}
         description={post.frontmatter.description}
@@ -78,16 +79,16 @@ const DesignGuidelinePost = ({ data }) => {
   )
 }
 
-DesignGuidelinePost.propTypes = {
+DeveloperGuidelinePost.propTypes = {
   data: PropTypes.shape({
     markdownRemark: PropTypes.object,
   }),
 }
 
-export default DesignGuidelinePost
+export default DeveloperGuidelinePost
 
 export const pageQuery = graphql`
-  query DesignGuidelinePostByID($id: String!) {
+  query DeveloperGuidelinePostByID($id: String!) {
     markdownRemark(id: { eq: $id }) {
       id
       html
